+++ examples/KWK_getcheckdata.py
@@ -22,6 +22,7 @@ retrieve_meas = False
 derive_stats = False
 plot_meas = False
 plot_stations = False
+write_stations_table = False
 
 start_date = pd.Timestamp(1870, 1, 1, tz="UTC+01:00")
 end_date = pd.Timestamp(2024, 1, 1, tz="UTC+01:00")
@@ -137,3 +138,13 @@ if plot_stations:
     fig, ax = kw.plot_stations(station_list=station_list_map, crs=28992, add_labels=False)
     fig.savefig(os.path.join(dir_base,'stations_map.png'), dpi=200)
 
+
+
+### WRITE CSV WITH STATION CODE/X/Y/EPSG
+if write_stations_table:
+    # TODO: consider making retrieve_catalog public
+    from kenmerkendewaarden.data_retrieve import retrieve_catalog
+    locs_meas_ts_all, _, _ = retrieve_catalog(crs=4326)
+    locs_ts = locs_meas_ts_all.loc[locs_meas_ts_all.index.isin(station_list)]
+    file_csv = os.path.join(dir_base, "station_locations.csv")
+    locs_ts[["Locatie_MessageID","X","Y","Coordinatenstelsel","Naam"]].to_csv(file_csv)