+++ kenmerkendewaarden/data_retrieve.py
@@ -429,6 +429,8 @@ def read_measurements(
         return ds_meas
 
     df_meas = xarray_to_hatyan(ds_meas)
+    # back to centimeters
+    df_meas['values'] *= 100
 
     if drop_duplicates:
         df_meas = drop_duplicate_times(df_meas)